Add files via upload
This was my first hands-on experience with web development. After I spent some time exploring the theoretical side of HTML and CSS, I headed to Visual Studio Code to deploy my acquired knowledge. The Website shows what I've learned in basic HTML, such as the basic structure of a website and the styling and positioning using CSS.